# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "version": "1.0.0",
    "title": "Swagger Petstore",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:1337",
  "basePath": "/",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/connection/find": {
      "get": {
        "tags": [
          "connection"
        ],
        "summary": "Get all Connection",
        "description": "",
        "operationId": "getConnection",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Connection"
            }
          },
          "400": {
            "description": "Invalid id supplied"
          },
          "404": {
            "description": "Connection not found"
          }
        }
      }
    },
    "/connection/update": {
      "post": {
        "tags": [
          "connection"
        ],
        "summary": "Update Connection",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateConnection",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Update Basestation object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Basestation"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/connection/delete": {
      "delete": {
        "tags": [
          "connection"
        ],
        "summary": "Delete Connection",
        "description": "",
        "operationId": "deleteConnection",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Connection id",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            },
            "headers": {
              "X-Rate-Limit": {
                "type": "integer",
                "format": "int32",
                "description": "calls per hour allowed by the user"
              },
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "date in UTC when token expires"
              }
            }
          },
          "400": {
            "description": "Invalid username/password supplied"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Basestation": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "lat": {
          "type": "string"
        },
        "lag": {
          "type": "string"
        },
        "bandwidth": {
          "type": "string"
        },
        "max_connection": {
          "type": "string"
        },
        "status_id": {
          "type": "integer",
          "format": "int32",
          "description": "Basestation Status"
        }
      },
      "xml": {
        "name": "Basestation"
      }
    },
    "Connection": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "address": {
          "type": "string"
        },
        "router_of": {
          "type": "string"
        },
        "router_brand": {
          "type": "string"
        },
        "router_model": {
          "type": "string"
        },
        "router_price": {
          "type": "string"
        },
        "drop_wire_of": {
          "type": "string"
        },
        "drop_wire_length": {
          "type": "string"
        },
        "price_per_meter": {
          "type": "string"
        },
        "is_wireless": {
          "type": "boolean"
        },
        "lat": {
          "type": "string"
        },
        "lag": {
          "type": "string"
        },
        "status_id": {
          "type": "integer",
          "format": "int32",
          "description": "Connection Status"
        },
        "customer_id": {
          "type": "number"
        },
        "basestation_id": {
          "type": "number"
        },
        "package_id": {
          "type": "number"
        },
        "salesman_id": {
          "type": "number"
        },
        "dealer_id": {
          "type": "number"
        }
      },
      "xml": {
        "name": "Connection"
      }
    },
...
  },
...
}